+++ x/gamm/pool-models/balancer/pool.go
@@ -10,7 +10,6 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 
-	"github.com/osmosis-labs/osmosis/v11/osmomath"
 	"github.com/osmosis-labs/osmosis/v11/x/gamm/pool-models/internal/cfmm_common"
 	"github.com/osmosis-labs/osmosis/v11/x/gamm/types"
 )
@@ -618,25 +617,11 @@ func (p Pool) SpotPrice(ctx sdk.Context, baseAsset, quoteAsset string) (spotPric
 		return sdk.Dec{}, errors.New("pool is misconfigured, got 0 weight")
 	}
 
-	defer func() {
-		// defer function to escape the panic when spot price overflows
-		if r := recover(); r != nil {
-			spotPrice = sdk.Dec{}
-			err = types.ErrSpotPriceInternal
-		}
-	}()
-
 	// spot_price = (Base_supply / Weight_base) / (Quote_supply / Weight_quote)
 	// spot_price = (weight_quote / weight_base) * (base_supply / quote_supply)
 	invWeightRatio := quote.Weight.ToDec().Quo(base.Weight.ToDec())
 	supplyRatio := base.Token.Amount.ToDec().Quo(quote.Token.Amount.ToDec())
-	fullRatio := supplyRatio.Mul(invWeightRatio)
-	// we want to round this to `SigFigs` of precision
-	spotPrice = osmomath.SigFigRound(fullRatio, types.SigFigs)
-	if spotPrice.GT(sdk.NewDec(2).Power(160)) {
-		spotPrice = sdk.Dec{}
-		err = types.ErrSpotPriceOverflow
-	}
+	spotPrice = supplyRatio.Mul(invWeightRatio)
 
 	return spotPrice, err
 }